Added a Bell Number calculator
The explosive growth in the number of ways to partition a set has important implications for finding communities in a network. A simple algorithm would be to check the average density of the subsets for every possible way to group the nodes into subsets, then select the configuration that maximizes the average density. However, this method becomes intractable with even small sizes of n. Thus we will have to find another way to search the solution space for communities within a network. And that, dear readers, will be a future adventure.